requirements/dev.txt
django==2.2.3
PyMySQL==0.8.0
mysqlclient
djangorestframework
xlrd==1.2.0
requests-html
pyecharts